move RV class tests to test_rv_class.py
