+++ tests/test_rv_utils.py
@@ -1,63 +1,14 @@
-import ephem
-import pytest
 import datetime
+
+import ephem
 import numpy as np
+import pytest
+from hypothesis import given
 from hypothesis import strategies as st
-from hypothesis import given, example, assume
 
 import rv
-from utils.rv_utils import RV, JulianDate
+from utils.rv_utils import JulianDate
 from utils.rv_utils import RV_from_params
-
-
-# issue with limits  0-pi only
-@given(st.lists(st.floats(min_value=0, max_value=np.pi), min_size=1), st.floats(min_value=0, max_value=1))
-def test_true_anomaly(ma, ecc):
-
-    ma = np.asarray(ma)
-    assume(np.all(np.abs(ma) > 0.0001))
-    ta = RV.true_anomaly(ma, ecc)
-    # cos(E) = (e + cos(ta)) / (1 + e*cos(ta))
-    E = np.arccos((ecc + np.cos(ta)) /(1 + ecc * np.cos(ta)))
-    # ma = E - e*sin(E)
-    print("VALUES", ma, (E - ecc * np.sin(E)))
-    assert np.allclose(ma, E - ecc * np.sin(E), rtol=0.05)
-    assert len(ta) == len(ma)
-
-
-# issue with limits 0-pi only
-@given(st.floats(min_value=0, max_value=np.pi), st.floats(min_value=0, max_value=1))
-@example(2, 0.5)   # example with an integer
-def test_true_anomaly_with_scalar(ma, ecc):
-    assume(abs(ma) > 0.001)
-    ta = RV.true_anomaly(ma, ecc)
-    # Contrast Result to E from ta
-    # cos(E) = (e + cos(ta)) / (1 + e*cos(ta))
-    E = np.arccos((ecc + np.cos(ta)) / (1 + ecc * np.cos(ta)))
-    # ma = E - e*sin(E)
-    MA = E - ecc * np.sin(E)
-    assert np.allclose(ma, E - ecc * np.sin(E))
-    assert len(ta) == 1
-
-
-@given(st.floats(min_value=0.01, max_value=0.99))
-def test_true_anomaly_errors(ecc):
-
-    with pytest.raises(TypeError):
-        RV.true_anomaly([], ecc)
-
-    with pytest.raises(ValueError):
-        RV.true_anomaly(np.array([]), ecc)
-
-
-@given(st.lists(st.floats(), min_size=1), st.floats(), st.floats(min_value=0.01))
-def test_mean_anomaly(t, t0, p):
-    """Mean anomaly is an angle, doesn't have a constraint value."""
-    t = np.array(t)
-    ma = RV.mean_anomaly(t, t0, p)
-
-    assert len(t) == len(ma)
-    assert isinstance(t, np.ndarray)
 
 
 @pytest.mark.xfail
